django/requirements.txt
Django
django-cors-headers

drf-spectacular
djangorestframework

djangorestframework-simplejwt


#drf-yasg

